make it work correctly

+++ backend/rest_api_service/crud.py
@@ -50,7 +50,8 @@ async def get_user_by_username(db: Session, username: str) -> Type[models.DBUser
         return None
 
 
-async def create_user(db: Session, username: str, email: str, password_hash: str, birthday: datetime.datetime) -> models.DBUser | None:
+async def create_user(db: Session, username: str, email: str, password_hash: str,
+                      birthday: datetime.datetime) -> models.DBUser | None:
     try:
         user = models.DBUser(
             username=username,
@@ -67,11 +68,11 @@ async def create_user(db: Session, username: str, email: str, password_hash: str
         db.refresh(user)
         return user
     except Exception:
-        db.rollback()  
-        return None 
+        db.rollback()
+        return None
 
 
-async def update_user_last_login(db: Session, id : int) -> None:
+async def update_user_last_login(db: Session, id: int) -> None:
     try:
         user = await get_user_by_id(db, id)
         if user:
@@ -79,8 +80,9 @@ async def update_user_last_login(db: Session, id : int) -> None:
             db.commit()
             db.refresh(user)
     except Exception:
-        db.rollback()  
+        db.rollback()
         return None
+
 
 async def get_event_by_id(event_id: int, db: Session):
     return db.get(models.DBEvent, event_id)
@@ -121,18 +123,22 @@ async def create_required_event_participants(db_event, db):
                                           response_time=datetime.datetime.now()))  # TODO: zgadnij kiedy odpowie
 
 
+async def create_event_categories(event_id: int, event: models.EventRequest, db: Session):
+    for category_id in event.categories:
+        db.add(models.DBEventCategory(event_id=event_id, category_id=category_id))
+
+
 async def create_event_with_required_associations(event: models.EventRequest, db: Session):
     db_event = models.DBEvent.create(event)
     db.add(db_event)
     db.commit()
     await create_required_event_participants(db_event=db_event, db=db)
-    for category_id in event.categories:
-        db.add(models.DBEventCategory(event_id=db_event.event_id, category_id=category_id))
+    await create_event_categories(event_id=db_event.event_id, event=event, db=db)
     db.commit()
 
 
 async def create_event_participant(event_participants: str, db: Session):
-    # TODO: consider is it useless and potentially remove
+    # TODO: consider is it useless and potentially remove (depends on future)
     db_event_participants = models.DBEventParticipants.create(event_participants)
     db.add(db_event_participants)
     db.commit()
@@ -148,7 +154,17 @@ async def update_event(event_id: int, changed_data: models.EventRequest, db: Ses
     event = db.get(models.DBEvent, event_id)
     new_data = changed_data.dict(exclude_unset=True)
     for field, value in new_data.items():
-        setattr(event, field, value)
+        if field == "categories":
+            await delete_all_event_categories(event_id=event_id, db=db)
+            await create_event_categories(event_id=event_id, event=changed_data, db=db)
+        else:
+            if field == "privacy" and not event.privacy == value:
+                await delete_all_event_participants(event_id=event_id, db=db)
+                setattr(event, field, value)
+                db.commit()
+                await create_required_event_participants(db_event=event, db=db)
+            else:
+                setattr(event, field, value)
     db.commit()
 
 
@@ -160,14 +176,46 @@ async def update_category(category_id: int, changed_data: models.CategoryRequest
     db.commit()
 
 
-async def delete_event(event_id: int, db: Session):
+async def delete_event_and_associated_objects(user_id: int, event_id: int, db: Session):
     event = db.get(models.DBEvent, event_id)
-    # TODO: delete all associations
-    db.delete(event)
+    if event.created_by == user_id:
+        await delete_all_event_categories(event_id=event_id, db=db)
+        await delete_all_event_participants(event_id=event_id, db=db)
+        db.delete(event)
+        db.commit()
+    else:
+        await accept_or_reject_participate_in_event(user_id=user_id, event_id=event_id, is_accepted=False, db=db)
+
+
+async def accept_or_reject_participate_in_event(user_id: int, event_id: int, is_accepted: bool, db: Session):
+    db_event_participant = db.query(models.DBEventParticipants).filter(
+        models.DBEventParticipants.event_id == event_id and models.DBEventParticipants.user_id == user_id).first()
+    if is_accepted:
+        db_event_participant.participant_status = "accepted"
+        db_event_participant.response_time = datetime.datetime.now()
+    else:
+        db.delete(db_event_participant)
+    db.commit()
+
+
+async def delete_all_event_participants(event_id: int, db: Session):
+    db_event_participants = db.query(models.DBEventParticipants).filter(
+        models.DBEventParticipants.event_id == event_id).all()
+    for db_event_participant in db_event_participants:
+        db.delete(db_event_participant)
+    db.commit()
+
+
+async def delete_all_event_categories(event_id: int, db: Session):
+    db_event_categories = db.query(models.DBEventCategory).filter(
+        models.DBEventCategory.event_id == event_id).all()
+    for db_event_category in db_event_categories:
+        db.delete(db_event_category)
     db.commit()
 
 
 async def delete_category(category_id: int, db: Session):
+    # TODO: Probably to remove
     category = db.get(models.DBCategory, category_id)
     db.delete(category)
     db.commit()
